base/requirements/base.txt
# https://github.com/joke2k/django-environ
django-environ

# ------------------------------------------------------------------------------
# Request
# ------------------------------------------------------------------------------
requests
# https://pynacl.readthedocs.io/en/latest/
PyNaCl

# https://docs.aiohttp.org/en/latest/index.html#aiohttp-installation
aiohttp
aiodns
Brotli
